+++ keys/Cardano/Config/Key/Praos.hs
@@ -9,7 +9,7 @@
 {-# LANGUAGE TypeFamilies #-}
 
 -- | Praos consensus key types and their 'Key' class instances
-module Cardano.Key.Praos
+module Cardano.Config.Key.Praos
   ( -- * Key types
     KesKey
   , VrfKey
@@ -26,19 +26,19 @@ module Cardano.Key.Praos
 where
 
 import Cardano.Binary.FixedSizeCodec qualified as Crypto
+import Cardano.Config.Key.Class
+import Cardano.Config.Key.HasTypeProxy
+import Cardano.Config.Key.Hash
+import Cardano.Config.Serialise.Cbor
+import Cardano.Config.Serialise.Raw
+import Cardano.Config.Serialise.TextEnvelope
+import Cardano.Config.Serialise.Using
 import Cardano.Crypto.DSIGN.Class qualified as Crypto
 import Cardano.Crypto.Hash.Class qualified as Crypto
 import Cardano.Crypto.KES.Class qualified as Crypto
 import Cardano.Crypto.VRF.Class qualified as Crypto
-import Cardano.Key.Class
-import Cardano.Key.HasTypeProxy
-import Cardano.Key.Hash
 import Cardano.Ledger.Hashes (HASH)
 import Cardano.Protocol.Crypto (KES, StandardCrypto, VRF)
-import Cardano.Serialise.Cbor
-import Cardano.Serialise.Raw
-import Cardano.Serialise.TextEnvelope
-import Cardano.Serialise.Using
 import Data.ByteString (ByteString)
 import Data.Either.Combinators (maybeToRight)
 import Data.String (IsString (..))